Done Store

+++ src/modules/brand/brand.service.ts
@@ -3,11 +3,14 @@ import RootService from '../root/root.service';
 import { BrandEntity } from './brand.entity';
 import { CreateBrandInput, UpdateBrandInput } from 'src/graphql.schema';
 import { PhotoService } from '../photo/photo.service';
+import { ProductService } from '../product/product.service';
+import { GraphQLError } from 'graphql';
 
 @Injectable()
 export class BrandService extends RootService<BrandEntity>{
   constructor(
-    private readonly photoService: PhotoService
+    private readonly photoService: PhotoService,
+    private readonly productService: ProductService
   ){
     super(BrandEntity, 'Thương hiệu')
   }
@@ -30,7 +33,11 @@ export class BrandService extends RootService<BrandEntity>{
   }
 
   async deleteBrands(ids: string[]): Promise<boolean>{
-    //TODO Check if brand is used
+    // Check if brand is used
+    const productByBrand = await this.productService.find({idBrand: {$in: ids}})
+    if(productByBrand.length)
+      throw new GraphQLError('Thương hiệu đã được sử dụng!')
+
     const existeds = await this.checkExistedIds(ids)
     existeds.forEach(brand => this.photoService.remove(brand.img))
     return this.delete(ids)
